+++ src/metaobject-repository.ts
@@ -4,21 +4,9 @@ import { FieldBuilder, QueryBuilder } from "raku-ql";
 import { camel } from "snake-camel";
 import { Collection, Company, Customer, GenericFile, Job, MediaImage, Metaobject, MetaobjectBulkDeletePayload, MetaobjectCreatePayload, MetaobjectDeletePayload, MetaobjectsCreatePayload, MetaobjectStatus, MetaobjectUpdatePayload, MetaobjectUpsertPayload, Page, PageInfo, Product, ProductVariant, TaxonomyValue, Video } from "~/types/admin.types";
 import { DefinitionSchema, DefinitionSchemaEntry, FieldDefinition, FromDefinitionWithSystemData, ValidPopulatePaths } from "./types/definitions";
-import { CreateInput, FindOptions, EmptyObjectOptions, OnPopulateFunc, PopulateOptions, SortKey, UpdateInput, UpsertInput } from "./types/metaobject-repository";
+import { CreateInput, FindOptions, EmptyObjectOptions, OnPopulateFunc, PopulateOptions, SortKey, UpdateInput, UpsertInput, EmptyObject } from "./types/metaobject-repository";
 import { UserErrorsException } from "./exception/user-errors-exception";
 import { deserialize, serializeFields } from "./transformer";
-
-type EmptyObject<T extends { system: object }> =
-  // 1) Remove the original `system`
-  Omit<T, "system"> & {
-    // 2) Re-add `system` with all keys except `capabilities` made nullable
-    system: {
-      [K in keyof T["system"]]:
-        K extends "type" | "capabilities"
-          ? T["system"][K]
-          : T["system"][K] | null;
-    };
-  };
 
 /**
  * Object repository
@@ -43,7 +31,7 @@ export class MetaobjectRepository<
    * Generate a new empty object that contains all the fields of the definition, set to empty value
    */
   
-  getEmptyObject(opts?: EmptyObjectOptions): EmptyObject<FromDefinitionWithSystemData<D, T>> {
+  getEmptyObject(opts?: EmptyObjectOptions<D, T>): EmptyObject<FromDefinitionWithSystemData<D, T>> {
     const definition = this.getDefinitionSchemaEntry(this.type);
 
     let data: any = {
@@ -75,7 +63,7 @@ export class MetaobjectRepository<
       const key = camel(field.key);
 
       if (opts?.defaultValues && key in opts.defaultValues) {
-        data[key] = opts.defaultValues?.[key];
+        data[key] = opts.defaultValues?.[key as keyof typeof opts.defaultValues];
       } else {
         data[key] = field.type.startsWith('list.') ? [] : null;
       }